refactor: Update StyledTitle imports and adjust props for consistency across components
