+++ components/promoter/PromoterArtistsClient.tsx
@@ -5,7 +5,7 @@ import { IconUser, IconMapPin, IconSearch, IconX, IconUserPlus, IconMusic, IconA
 import Link from "next/link";
 import { nameToUrl } from "@/lib/utils";
 import { useState } from "react";
-import { StyledTitle } from "../StyledTitle/StyledTitle";
+import StyledTitle from "@/components/StyledTitle";
 
 interface Artist {
   id: string;
@@ -238,14 +238,16 @@ function ArtistCard({
         {/* Artist Info */}
         <Stack gap="xs" align="center" style={{ flex: 1 }}>
           <StyledTitle 
-            title={artist.name}
-            fontName="Inter"
-            order={4}
-            fw={700} 
-            size="lg" 
-            ta="center" 
-            style={{ lineClamp: 1 }}
-          />
+            selectedFont="Inter"
+            style={{ 
+              lineClamp: 1, 
+              fontWeight: 700, 
+              fontSize: '1.125rem',
+              textAlign: 'center' 
+            }}
+          >
+            {artist.name}
+          </StyledTitle>
           
           {/* Location */}
           <Group gap="xs" justify="center">
